+++ tests/test_simplified_procedure.f90
@@ -13,8 +13,8 @@ program test_simplified_procedure
     implicit none
 
     ! Variables
-    real :: M, ntot, press, T, rho
-    real, dimension(NUM_SP) :: y, x
+    real(8) :: M, ntot, press, T, rho
+    real(8), dimension(NUM_SP) :: y, x
     integer :: i, j, k, n
     type(transport_in) :: transport
     type(transport_out) :: transport_coeff
